rename Fun4AllPrdfInputCombinerManager to Fun4AllPrdfInputPoolManager
