+++ offline/framework/fun4allraw/Fun4AllPrdfInputPoolManager.h
@@ -5,11 +5,15 @@
 
 #include <fun4all/Fun4AllInputManager.h>
 
+#include <Event/phenixTypes.h>
+
 #include <map>
 #include <string>
 
 class Event;
-class Eventiterator;
+class SinglePrdfInput;
+class oEvent;
+class Packet;
 class PHCompositeNode;
 class SyncObject;
 
@@ -18,8 +22,8 @@ class Fun4AllPrdfInputPoolManager : public Fun4AllInputManager
  public:
   Fun4AllPrdfInputPoolManager(const std::string &name = "DUMMY", const std::string &prdfnodename = "PRDF", const std::string &topnodename = "TOP");
   ~Fun4AllPrdfInputPoolManager() override;
-  int fileopen(const std::string &filenam) override;
-//cppcheck-suppress virtualCallInConstructor
+  int fileopen(const std::string &filenam) override { return 0; }
+  // cppcheck-suppress virtualCallInConstructor
   int fileclose() override;
   int run(const int nevents = 0) override;
 
@@ -28,21 +32,29 @@ class Fun4AllPrdfInputPoolManager : public Fun4AllInputManager
   int PushBackEvents(const int i) override;
   int GetSyncObject(SyncObject **mastersync) override;
   int SyncIt(const SyncObject *mastersync) override;
-  int HasSyncObject() const  override {return 1;}
+  int HasSyncObject() const override { return 1; }
   std::string GetString(const std::string &what) const override;
+  //  Eventiterator *AddPrdfInputList(const std::string &listfile);
+  SinglePrdfInput *AddPrdfInputFile(const std::string &filename);
+  void AddPacket(const int evtno, Packet *p);
+  void UpdateEventFoundCounter(const int evtno);
 
  private:
-  int m_Segment = -999;
-  int m_EventsTotal = 0;
-  int m_EventsThisFile = 0;
-  unsigned int m_PoolDepth = 10;
-  unsigned int m_LowWaterMark = 5;
+  struct PacketInfo
+  {
+    std::vector<Packet *> PacketVector;
+    unsigned int EventFoundCounter = 0;
+  };
+
+  int m_RunNumber = 0;
+  std::vector<SinglePrdfInput *> m_PrdfInputVector;
+  SyncObject *m_SyncObject = nullptr;
   PHCompositeNode *m_topNode = nullptr;
   Event *m_Event = nullptr;
-  Event *m_SaveEvent = nullptr;
-  Eventiterator *m_EventIterator = nullptr;
-  SyncObject *m_SyncObject = nullptr;
-  std::map<int, Event *> m_EventPool;
+  PHDWORD workmem[4 * 1024 * 1024] = {};
+  oEvent *oph;
+
+  std::map<int, PacketInfo> m_PacketMap;
   std::string m_PrdfNodeName;
 };
