+++ offline/framework/fun4allraw/Fun4AllRolloverFileOutStream.cc
@@ -31,7 +31,7 @@ Fun4AllRolloverFileOutStream::Fun4AllRolloverFileOutStream(const std::string &fr
   {
     if (m_MaxFileFize > MaxSize())
     {
-      unsigned long long maxmb = MaxSize() / (1024 * 1024);
+      unsigned long long maxmb = MaxSize() / (1024ULL * 1024ULL);
       std::cout << "setting maximum size to current max (in MB): " << maxmb << std::endl;
     }
     m_MaxFileFize = MaxSize();
