+++ offline/framework/fun4allraw/Makefile.am
@@ -15,7 +15,6 @@ pkginclude_HEADERS = \
   Fun4AllEventOutputManager.h \
   Fun4AllFileOutStream.h \
   Fun4AllPrdfInputManager.h \
-  Fun4AllPrdfInputCombinerManager.h \
   Fun4AllPrdfInputPoolManager.h \
   Fun4AllPrdfOutputManager.h \
   Fun4AllRolloverFileOutStream.h \
@@ -29,7 +28,6 @@ libfun4allraw_la_SOURCES = \
   Fun4AllEventOutputManager.cc \
   Fun4AllFileOutStream.cc \
   Fun4AllPrdfInputManager.cc \
-  Fun4AllPrdfInputCombinerManager.cc \
   Fun4AllPrdfInputPoolManager.cc \
   Fun4AllPrdfOutputManager.cc \
   Fun4AllRolloverFileOutStream.cc \
